chore(build): made dependencies devDependencies
Moved all the webpack dependencies to devDependencies.

When including the plugin in projectes who do not use stylus
the dependencies produce problems.
The `stylus-loader` for example has a peerDependency for stylus
which might not be part of the project. Then `npm list` fails
and serveral build-steps do not work anymore.